requirements.txt
python-pptx
Pillow
requests
open_clip_torch
transformers
torch
openai
comtypes; platform_system == "Windows"
psutil
ollama
accelerate
pytictoc
flash_attn; platform_system == "Windows"
mlx_vlm; platform_system == "Darwin"
